Destination infrastructure created in DestinationPickerBottomSheetDialog (VerticalMarginItemDecoration, row_item_divider.xml, DestinationModel, DestinationsAdapter, destination_item_layout.xml)

+++ utils/src/main/java/com/livmas/utils/dpHelper.kt
@@ -2,10 +2,10 @@ package com.livmas.utils
 
 import android.content.Context
 
-fun Context.dpToPx(dp: Int): Int {
+fun Context.dpToPx(dp: Float): Int {
     return (dp * resources.displayMetrics.density).toInt()
 }
 
-fun Context.pxToDp(px: Int): Int {
-    return (px / resources.displayMetrics.density).toInt()
+fun Context.pxToDp(px: Int): Float {
+    return px / resources.displayMetrics.density
 }
